docs: multiple instance docs
