+++ examples/configuration/multipleInstance/admin-v1-docs.js
@@ -1,0 +1,5 @@
+/**
+ * GET /api/admin/v1
+ * @summary This is the summary or description of the endpoint
+ * @return {string} 200 - success response
+ */
