+++ examples/configuration/multipleInstance/client-v1-docs.js
@@ -1,0 +1,5 @@
+/**
+ * GET /api/client/v1
+ * @summary This is the summary or description of the endpoint
+ * @return {string} 200 - success response
+ */
