+++ examples/package.json
@@ -6,6 +6,7 @@
   "scripts": {
     "configuration:main": "node configuration/main.js",
     "configuration:multipleFiles": "node configuration/multipleFiles.js",
+    "configuration:multipleInstance": "node configuration/multipleInstance/index.js",
     "parameters:simple": "node parameters/simple.js",
     "parameters:headers": "node parameters/headers.js",
     "parameters:multiple": "node parameters/multiple.js",
